requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
openai>=1.80.1
python-dotenv>=1.0.1
tiktoken>=0.5.0
